requirements.txt
fastapi==0.110.0
uvicorn>=0.31.1
crewai>=1.12.0
langchain-groq
python-dotenv
setuptools<70.0.0